+++ flairsou/urls.py
@@ -22,6 +22,7 @@ from drf_spectacular.views import SpectacularAPIView, \
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('api/', include('flairsou_api.urls')),
+    path('proxy_portail/', include('proxy_portail.urls')),
     path('mock_portail/', include('mock_portail.urls')),
     path('doc/schema/', SpectacularAPIView.as_view(), name='schema'),
     path('doc/schema/swagger-ui/',
