+++ flairsou_api/views/__init__.py
@@ -5,9 +5,6 @@ from .book_views import BookAccountList
 from .transaction_views import OperationDetail, OperationList
 from .transaction_views import TransactionDetail, TransactionList
 from .reconciliation_views import ReconciliationView
-from .user_views import get_authorization_link, request_oauth_token
-from .user_views import get_user_infos
-from .user_views import user_logout
 
 __all__ = [
     AccountDetail,
@@ -23,8 +20,4 @@ __all__ = [
     TransactionDetail,
     TransactionList,
     ReconciliationView,
-    get_authorization_link,
-    request_oauth_token,
-    get_user_infos,
-    user_logout,
 ]
