+++ proxy_portail/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class ProxyPortailConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'proxy_portail'
