+++ proxy_portail/views/__init__.py
@@ -1,0 +1,12 @@
+from .oauth_views import get_authorization_link
+from .oauth_views import request_oauth_token
+from .oauth_views import user_logout
+
+from .user_views import get_user_infos
+
+__all__ = [
+    get_authorization_link,
+    request_oauth_token,
+    user_logout,
+    get_user_infos,
+]
